# docs/openapi.slice1.yaml
openapi: 3.1.0
info:
  title: Heard API
  version: 0.5.0
  description: >
    Heard's tenant-scoped campaign, guest feedback, and recovery APIs. Manager
    routes require a Passage-issued Bearer product token and matching tenant membership.
    Production validation requires EdDSA with Passage JWKS, the configured issuer,
    audience `heard`, token expiry, UUID subject and organization claims, a known
    organization role, and a `heard` product-access grant. Client role or tenant
    headers never create identity authority.
servers:
  - url: http://localhost:3010
    description: Same-origin web gateway for local development
security:
  - bearerAuth: []
paths:
...
  /api/v1/marketing-leads:
    post:
      summary: Request a heard product conversation
      description: Stores a restaurant prospect's explicit request to be contacted by heard.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMarketingLeadRequest"
      responses:
        "201":
          description: Contact request saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketingLead"
        "400":
          description: Contact request is incomplete or malformed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/auth/local/session:
    post:
      summary: Issue a local Passage-compatible development session
      description: Available only when the non-production local Passage adapter is enabled.
      security: []
      x-local-development-only: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLocalSessionRequest"
      responses:
        "201":
          description: Local development session issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/auth/local/registration:
    post:
      summary: Register through the local Passage-compatible development adapter
      description: >
        Creates a predictable non-production Passage account context for the
        self-service onboarding workflow. This endpoint is unavailable in
        production and is not the production Passage registration contract.
      security: []
      x-local-development-only: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLocalSessionRequest"
      responses:
        "201":
          description: Local development registration resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/session:
    get:
      summary: Resolve the current Passage identity
      responses:
        "200":
          description: Authenticated identity and claims
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Identity"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/onboarding:
    get:
      summary: Resolve the current operator's Heard onboarding state
      responses:
        "200":
          description: Current durable onboarding state and next incomplete step
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingState"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/onboarding/activations:
    post:
      summary: Create or resume the first restaurant workspace and location
      description: >
        Derives actor and account context from the verified Passage identity.
        Retries and concurrent requests for the same actor return the same
        activation, tenant, and location.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOnboardingActivationRequest"
      responses:
        "200":
          description: Workspace activation created or safely resumed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingState"
        "400":
          description: Activation data or account context is incomplete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Verified identity may not activate a Heard workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/tenants:
    post:
      summary: Create tenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTenantRequest"
      responses:
        "201":
          description: Tenant created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
  /api/v1/tenants/{id}:
    get:
      summary: Get tenant
      parameters:
        - $ref: "#/components/parameters/TenantId"
      responses:
        "200":
          description: Tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
  /api/v1/locations:
    get:
      summary: List locations for the active tenant
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Location collection
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Location"
                  page: { type: integer }
                  page_size: { type: integer }
                  has_more: { type: boolean }
    post:
      summary: Create location
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLocationRequest"
      responses:
        "201":
          description: Location created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Location"
  /api/v1/survey-campaigns:
    get:
      summary: List survey campaigns for the active tenant
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Survey campaign collection
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/SurveyCampaign"
                  page: { type: integer }
                  page_size: { type: integer }
                  has_more: { type: boolean }
    post:
      summary: Create a survey campaign
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSurveyCampaignRequest"
      responses:
        "201":
          description: Survey campaign created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SurveyCampaign"
  /api/v1/survey-campaigns/{id}:
    get:
      summary: Get a survey campaign
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Survey campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SurveyCampaign"
  /api/v1/public/surveys/{token}:
    get:
      summary: Resolve a public survey campaign by opaque token
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Public survey and feedback link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicSurvey"
  /api/v1/locations/{id}:
    get:
      summary: Get location
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - $ref: "#/components/parameters/LocationId"
      responses:
        "200":
          description: Location
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Location"
  /api/v1/feedback-links:
    post:
      summary: Create feedback link
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFeedbackLinkRequest"
      responses:
        "201":
          description: Feedback link created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackLink"
  /api/v1/feedback-links/{id}/qr:
    post:
      summary: Regenerate QR for feedback link
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - $ref: "#/components/parameters/FeedbackLinkId"
      responses:
        "200":
          description: Feedback link with QR payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackLink"
  /api/v1/feedback-links/resolve/{token}:
    get:
      summary: Resolve feedback link token
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Resolved feedback link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackLink"
  /api/v1/feedback-sessions:
    post:
      summary: Create feedback session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFeedbackSessionRequest"
      responses:
        "201":
          description: Feedback session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackSession"
  /api/v1/feedback-responses:
    get:
      summary: List feedback responses
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Feedback response collection
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/FeedbackResponse"
                  page: { type: integer }
                  page_size: { type: integer }
                  has_more: { type: boolean }
    post:
      summary: Submit feedback response
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitFeedbackRequest"
      responses:
        "201":
          description: Feedback response created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackResponse"
  /api/v1/feedback-responses/{id}:
    get:
      summary: Get feedback response
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Feedback response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackResponse"
  /api/v1/recovery-cases:
    get:
      summary: List recovery cases
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Recovery case collection
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/RecoveryCase"
                  page: { type: integer }
                  page_size: { type: integer }
                  has_more: { type: boolean }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Passage access token
...
  parameters:
...
    TenantId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [message, status]
          properties:
            message: { type: string }
            status: { type: integer }
...
    CreateLocalSessionRequest:
      type: object
      required: [email]
      additionalProperties: false
      properties:
        email: { type: string, format: email }
...
    CreateMarketingLeadRequest:
      type: object
      required: [name, work_email, restaurant_name, location_count, contact_consent]
      additionalProperties: false
      properties:
        name: { type: string, minLength: 2, maxLength: 120 }
        work_email: { type: string, format: email, maxLength: 254 }
        phone: { type: string, maxLength: 32, pattern: "^[+()0-9 .-]*$" }
        restaurant_name: { type: string, minLength: 2, maxLength: 160 }
        location_count: { type: string, enum: ["1", "2-4", "5-19", "20-49", "50+"] }
        challenge: { type: string, maxLength: 1000 }
        source: { type: string, enum: [marketing_site, guest_demo], default: marketing_site }
        contact_consent: { type: boolean, const: true }
...